style: Organize file places
